// src/services/@drazisil/mco-types/types.ts
export interface IRawPacket {
  connectionId: string
  connection: TCPConnection
  data: Buffer
  localPort: number | undefined
  remoteAddress: string | undefined
  timestamp: number
}